requirements.txt
# RAG_Rice Project Dependencies

# Core ML/AI Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-session>=0.5.0
gunicorn>=21.2.0

# LLM and RAG
groq>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10

# Vector Database
chromadb>=0.4.20

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.35.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web Framework (if using Flask/FastAPI)
flask>=3.0.0
flask-cors>=4.0.0
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Data Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
pylint>=3.0.0

# Jupyter (for development)
jupyter>=1.0.0
ipykernel>=6.26.0

# Additional ML Tools
joblib>=1.3.0
scipy>=1.11.0

# Type Hints
typing-extensions>=4.8.0

# Progress Tracking
rich>=13.7.0
